+++ frontend/src/sidepanel/components/Quiz/Feedback.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+
+export const Feedback = ({
+  feedback,
+  isFeedbackLoading,
+}: {
+  feedback: string | null;
+  isFeedbackLoading: boolean;
+}) => {
+  if (isFeedbackLoading) {
+    return <div>解答を採点中...</div>;
+  }
+  if (feedback == null) {
+    return <></>;
+  }
+  return <div>採点結果：{feedback}</div>;
+};
