+++ frontend/src/sidepanel/components/Quiz/index.ts
@@ -1,0 +1,4 @@
+import { Quiz, QuizCard } from "./Quiz";
+
+export { QuizCard };
+export type { Quiz };
